refactor: improve two column layout component
Allow key to be specified on items
Add propTypes
Use `children` prop instead of `items`
