+++ app/components/global/index.js
@@ -1,5 +1,3 @@
 export { default as AppBar } from './AppBar'
-export {
-  default as AuthenticatedComponent,
-} from './AuthenticatedComponent'
-export { default as Layout } from './Layout'
+export { default as AuthenticatedComponent } from './AuthenticatedComponent'
+export { default as Layout } from './layout/Layout'
