+++ app/components/global/layout/TwoColumnLayout.md
@@ -1,0 +1,10 @@
+Responsive layout with two columns on wider screens and a single column on
+smaller screens.
+
+```js
+;<TwoColumnLayout>
+  <h1>Some thing</h1>
+  <h1>Something else</h1>
+  <h1>Another thing</h1>
+</TwoColumnLayout>
+```
